feat(events): move PDF generation button to Events page
Closes #221.

The race-results PDF dropdown (organiser summary / podium / racer
certificates) is an event-level action — it summarises all races within
an event, not a single race. Move it from the Race Admin page to the
Events page where it belongs UX-wise.

Behaviour:
- PDF button only enabled when exactly one event is selected in the
  events table.
- Loading state aggregated across the three PDF types for the selected
  event, same pattern as before.
- All other Race Admin actions (export CSV, edit, delete races) stay
  on the Race Admin page where they belong.

No backend changes — the same `usePdfApi` hook is used from the new
location. No unit test — this is a UI placement change and the
underlying mutation flow is unchanged.
